Move tests to unit and integration packages
