+++ src/test/java/ch/wisv/areafiftylan/integration/XAuthIntegrationTest.java
@@ -15,8 +15,9 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ch.wisv.areafiftylan;
+package ch.wisv.areafiftylan.integration;
 
+import ch.wisv.areafiftylan.ApplicationTest;
 import ch.wisv.areafiftylan.exception.TicketOptionNotFoundException;
 import ch.wisv.areafiftylan.products.model.Ticket;
 import ch.wisv.areafiftylan.products.model.TicketOption;
